Move unixhttp classes to 'haproxy.py' collector
